build(windows): fix the 34k clang-cl warning flood + a real long-overflow bug
The Windows CI job (clang-cl) emitted ~34k warnings dominated by -Weverything-
tier flags (unsafe-buffer-usage, old-style-cast, sign-conversion, …). Root cause:
clang-cl maps the GCC-style `-Wall` we pass to `-Weverything`. vgre_set_compiler_
warnings now detects clang-cl (Clang id + MSVC simulate id) and gives it the same
curated /W3 set as real MSVC instead of -Wall, so the log carries only meaningful
warnings. Linux/macOS (real GNU/Clang) keep -Wall -Wextra -Wpedantic -Werror.

Real bugs fixed while here:
- config_registry: registerInt/Spec used `long` (32-bit on Windows), so
  VGRE_SHM_RESULT_OFFSET's max 1099511627776 (2^40) silently truncated to 0 and
  std::stol would truncate a valid large value. Widened to `long long` + std::stoll
  so the 64-bit range is honored on every platform.
- Define _WINSOCK_DEPRECATED_NO_WARNINGS on Windows (alongside _CRT_SECURE_NO_WARNINGS)
  to silence the inet_addr-style deprecation notices the winsock headers emit.

Linux build unaffected (config change is a wider int type; the CMake changes are
WIN32/clang-cl-only).

+++ include/vgre/common/config_registry.h
@@ -23,7 +23,9 @@ public:
     static ConfigRegistry &instance();
 
     // Register the known knobs. `secret` masks the value in logs.
-    void registerInt(const std::string &name, long lo, long hi, long def,
+    // 64-bit range: on Windows `long` is 32-bit, which truncates large bounds
+    // (e.g. VGRE_SHM_RESULT_OFFSET's 2^40) — use `long long` so every platform agrees.
+    void registerInt(const std::string &name, long long lo, long long hi, long long def,
                      const std::string &desc);
     void registerEnum(const std::string &name, std::vector<std::string> values,
                       const std::string &def, const std::string &desc);
@@ -42,7 +44,7 @@ private:
     struct Spec {
         std::string name, desc, def;
         Kind kind;
-        long lo = 0, hi = 0;
+        long long lo = 0, hi = 0;
         std::vector<std::string> values;  // for Enum
         bool secret = false;
     };
